NavView now changes
Now NavigationView changes according to FirebaseAuth.getCurrentUser()
If a User i logged in when the MapActivity is created, then it clears the navigationMenu, and then it infaltes the other Menu.

For some reason I cannot add anymore than 4 items to a Menu. If I do so, it crashes the app. I do not know why, and it is quite annoying...
I would have added a Sign out button as well. That might be possible to just add an ImageButton like the Close Menu button.