AccountActivity under development
I am currently making the AccountActivity which is gonna show your accounts information. If you are a Facebook user it will show the Name, and if it is a Firebase User it will shwo the email.
Then there will be severeal other fields, and if a TextView is empty, then it should show red. And you should be able to click any of the TextViews and get a Dialog asking to enter the new value and then two buttons, one for Ok and one for Cancel.

Maybe there will be a button which then sends the user to the users selling location, and if he does not have one, it will ask if he wants to create one.
It will determine if the user has one with the new variable in firebase userInfo which is an Integer that holds how many sellingLocations the user currently has.