AccountActivity Done
AccountActivity should be done now.
It shows the users information in some TextViews, if the user does not have any value in the field, ex. name, then it says NoName and the text is red and clickable. You can then click the red text, and a dialog pops up asking for the new value. when the user pushes the change button, it then asks if the user are totally sure, as it cannot be changed again, if yes it changes it, if no it does nothing.