Change language on click.
It is now possible ot change ot the disired language with the help if the Spinner in the Title_bar.
At default it has a picture af the 5 different flags. And if you choose one, the background will be the flag corresponding to the selected language.

If the user selects his own language the Language name ex. "DA", will be saved in a shared preference so all the other Activities can access which language the user has selected.  AND if the user changes language in another Activity but the Maps-, then the MapsActivity will be able to know what the user has selected as well. The intent.putString cannot work on the MainActivity as the getString will not work and will always crash the app. But this way the MapsActivity knows what the user selected the last time he slected something, and will change to that language.