CreateLocationActivity under development
I am in the proccess of making the filter dialog box. I have made it so when you choose Berries the second spinner will be filled with berries. And so on.
Now I am making the button listener, so if the user clicks the ok button it checks if they have chosen anything. If not it has to say error. If they have selected anything, we will set the overallCategory or 2/3/4 or 5. And we will set specificItem1 or 2/3/4 or 5. If they are not the same.