CreateLocationActivity under development
It now takes the specifcItems and overallCateogry's from the popup box. And it changes the picture below to what the user has selected.
When the suer clicks create, it will save the SellingLocation under rootSellingLocations, specificSellingLocations and overallSellingLocations.

And the users data will be updated. So that the locationID under the user will be correct/added. And the users "numberOdCreatedLocations" will be updated to 1. As he now has one.
Users will  only be ablt to go under the createSellingLocations if the users "numberOfCreatedLocations" = 0. If it is 1. then he will be asked to go under change.