CreateLocationActivity under development
The picture function now works. When you press the "+" button it opens the camera. then you can take a picture and click ok. Then it puts the picture in the imageview and saves it in a bitmap for when the user want's to create the selling location.

The user fills out the rest of the form, and presses the "Create" button, it then proceeds to upload the picture to firebase storage, and saves the downloadURL and photoID in some variables and then it creates the selling location with all the variables such as the RoadName and No., zipcode, city, userID of creator, photoID, photoDownloadURL, overallCategory1/2/3/4/5 and SpecificItem1/2/3/4/5.

The only thing that is missing from this now is an if-else statement checking whether photoID and photoDownloadURl are empty or not.
And the button for getting your current location.