CreateLocationActivity almost done
The button witht eh arrow now works. It's function is that the address fields will be filled out automatically. The only issue with this is that the number of the address is in the same variable as the road. So when clicking the No. will be in the road field. The user will have to delete that and write it in the number field. The number is sometimes wrong anyways, I don't knwo why, I might put down the marker before getting the right location. Most of the time it is almost right, only the house number is wrong.