CreateLocationActivity Done + SortingController
Now is the CreateLocationActivity done, only missing a couple of small features like, when long pressing a specificItemIcon it should make it dissapere.

The SortingController is also done. Meaning that you can now whether you are signed in or not, press the filter button in the top and select what you want to see.